feat(docs): rewrite docs for datepicker in TS
Rewrites the documentation for the datepicker components in the
`src/pages/components/datepicker` folder in TypeScript.

All the changes are straightforward.

Explicitly imports and registers Buefy components so that they are type
checked. Note that no type checking is performed for globally registered
components. If there is no `<script>` block in a component file, we can
simply add a `<script setup lang="ts">` block and just import necessary
Buefy components in it.