test(lib): rewrite navbar spec in TS
Rewrites the specs for the navbar components in the
`src/components/navbar` in TypeScript.

In `NavBarItem.spec.js→ts`, mocks the `closeMenu` method of the
`stubNavBar` component instead of that of `wrapper.vm.$parent` to avoid
nasty type casting.

Applies the `vitest` package:
- Imports the spec building blocks from the `vitest` package
- Replaces `jest` with `vi` provided by the `vitest` package

In the `__snapshots__` subfolder, replaces the spec snapshots produced
by Jest with those by Vitest:
- `NavBar.spec.js.snap` → `NavBar.spec.ts.snap`
- `NavBarBurger.spec.js.snap` → `NavBarBurger.spec.ts.snap`
- `NavBarItem.spec.js.snap` → `NavBarItem.spec.ts.snap`
- `NavbarDropdown.spec.js.snap` → `NavbarDropdown.spec.ts.snap`