feat(docs): rewrites navbar docs in TS
Rewrites the documentation for the navbar components in the
`src/pages/components/navbar` in TypeScript.

All the changes are straightforward.

Appends a new `<script setup lang="ts"> block to `examples/ExSimple.vue`
and imports other Buefy components so that they are type checked.  Note
that no type checking is performed for globally registered components.