feat(docs): rewrite docs for tooltip in TS
Rewrites the documentation for `Tooltip` in the
`src/pages/components/tooltip` folder in TypeScript.

In most of the example components where no `<script>` block exists, adds
a new `<script setup lang="ts">` block and imports necessary Buefy
components so that they are type-checked.

In `examples/ExToggle.vue`, explicitly imports and register necessary
Buefy components so that they are type-checked.

Note that no type-checking is performed for globally registered
components.