feat(lib): rewrite utils/color in TS
Rewrites `src/utils/color.js` in TypeScript.

Unrolls the loops to dynamically define the following Color components
as pairs of `get` and `set` methods:
- `red`
- `green`
- `blue`
- `alpha`
- `hue`
- `saturation`
- `lightness`

My attempt to augment the `Color` class with an ambient module did not
work when all the Buefy types were bundled in a single `.dts` file;
i.e., it won't work in users' environments.

Replaces `get [Symbol.toString]` with `get [Symbol.toStringTag]` to
suppress a type error, although, I am not 100% confident about this
solutation because I was not able to identify the intention of the
original code.

`color.ts` exports color model types: `Hsl`, `Hsla`, `Rgb`, and `Rgba`