feat(lib): rewrite utils/color in TS
Rewrites `src/utils/color.js` in TypeScript.

The `Color` class dynamically defines some of its properties (getters
and setters of the color component properties; e.g., "alpha"), and I was
not able to figure out how to simply declare these types in the class
body. My circumvention was to augment the `Color` class with an ambient
module defined in another module:
- Renames `color.js` → `_color.ts` and exports the `Color` class with
  the name; i.e., non-default export
- Introduces a brand new `color.ts` which imports `Color` from
  `_color.ts`, augments it with the dynamically defined (but finite)
  properties, and reexports it as the default

Replaces `get [Symbol.toString]` with `get [Symbol.toStringTag]` to
suppress a type error, although, I am not 100% confident about this
solutation because I was not able to identify the intention of the
original code.

`color.ts` exports color model types: `Hsl`, `Hsla`, `Rgb`, and `Rgba`