chore(docs): analyze dependency tiers
Configures Vite so that `rollup-plugin-visualizer` analyzes the
dependencies and outputs the results in `tiers/stats.json`.

Introduces a new build script `build/analyze-tiers.mjs` which analyzes
dependency tiers from the output of `rollup-plugin-visualizer` and
outputs the results in `tiers/tiers.json`. The script is almost the
clone of the one in the `/packages/buefy-next/build` folder.

Introduces the following npm scripts:
- `analyze-tiers`: runs `build/analyze-tiers.mjs`
- `build:tiers`: runs `rollup-plugin-visualizer` (`build-only`) and then
  runs `analyze-tiers`. Use this script to make sure you have the latest
  dependency analysis results.