chore(docs): install @types/sortablejs
Adds `@types/sortablejs` to the dev dependencies so that the example of
the sortablejs extension can be type checked.

I have no idea about which version of `@types/sortablejs` I should
install, though, it should not matter as long as the code actually using
`sortablejs` does not change, and type-checking passes.