feat(gha): add workflow for lint, type-check, and unit test
Introduces a new GitHub Actions workflow `verify.yml` which performs
linting, type-checking, and unit testing when a commit is pushed or a
pull request is made to a specific branch. It will replace
`unit_testing.yml` when the TypeScript migration is complete.

Although the `migrate-to-typescript` is included in the trigger
branches, it should be removed when the TypeScript migration is
complete.