feat(lib): remove FormElementMixin's autocomplete & revert Taginput
Removes the `autocomplete` prop from `FormElementMixin`, and reverts the
breaking change that altered the `autocomplete` prop type of `Taginput`
from `Boolean` to `String`. The following components which mix
`FormElementMixin` in define their own `autocomplete` prop:
- `Autocomplete`
- `Input`
- `Numberinput`

Background: I introduced the breaking change to `Taginput` at the commit
085bd8583d3da221673261b1287d97bbb7d80fa1. The change was intended to
resolve the discrepancy in the `autocomplete` prop between `Taginput`
and `FormElementMixin`, which ended up with weird type errors. I should
not have changed the `autocomplete` prop type of `Taginput` but should
have removed the prop from `FormElementMixin` instead, because
`FormElementMixin` itself did not use it but just defined it; it was
there only to save the typing cost in host components which use
`FormElementMixin`, and most of the host components do not even use the
`autocomplete` prop.