fix(lib): Numberinput validation not work with Vue 3.4.28↑
Fixes the issue that a user was able to input an invalid number to
`Numberinput`. The underlying `Input` uses `v-model` instead of the
combination of the `:value` binding and the `@input` and `@change` event
handlers to work around the upstream issue introduced by Vue 3.4.28.
(See the background) The `onInput` and `onChange` methods of `Input` no
longer update the model value but just run validation.

Due to the limitation of `v-model`, almost the same `<input>` and
`<textarea>` elements has to be repeated to implement the variants of
the `lazy` prop. (See the background)

Note that this commit will cause type errors on the `v-model` bindings
of `<textarea>`. The error will be fixed with Vue 3.4.6 or higher.

Background:

`Numberinput` should reject numbers which are not equal to the basis for
stepping, and the basis is derived from the `value` attribute unless the
`min` is given. Before Vue 3.4.28, updating the `value` prop did not
affect the `value` attribute, however, Vue 3.4.28 or higher introduced a
tweak that updates the `value` attribute when the `value` prop is
changed. Since the `value` attribute which defines the basis moves, the
validation of `Numberinput` can be messed up. The problematic tweak is
not applied to `v-model` which the underlying `Input` did not use.

The reason why `Input` did not use `v-model` was to deal with the `lazy`
prop. While `v-model` needs an optional `.lazy` modifier if the `lazy`
prop is `true`, unfortunately, we cannot dynamically turn on / off the
`.lazy` modifier on `v-model`. (`modelModifiers` prop won't work with
`.lazy` modifier`.) This is why I had to duplicate the `<input>` and
`<textarea>` elements.