fix(docs): type errors in a fresh clone
Fixes the issue that a bunch of type errors similar to the following
were produced in a fresh clone:

    Error: src/App.vue(33,22): error TS2339: Property '$buefy' does not exist on type 'CreateComponentPublicInstance<{}, {}, { clipboard: ClipboardJS | null; }, {}, { setupClipboardControls(): void; }, ComponentOptionsMixin, ComponentOptionsMixin, ... 12 more ..., {}>'.

While I made a change to utilize the Project Reference feature of
TypeScript in the previous commit
8b513e7a1966bad1f21edee7f00b62b08c5f33d4, it required to build the
`buefy-next` package prior to the `docs` project. This is inconvenient
for the development purpose. I decided to give up the Project Reference
feature and directly include source files in the `../buefy-next/src`
folder. As the `../buefy-next/src` folder is included in `rootDirs` this
time, it is no longer illegal to import files from there.