feat(lib): make Clockpicker input lazy
Makes the underlying `Input` of `Clockpicker` lazy so that the input
text is not prematurely parsed. This change simplifies Moritz-Schmidt's
approach that introduced the data field `localInputValue` to temporarily
store the input.

According to my observation, the `@change` handler of `Clockpicker` was
called before that of the underlying `Input` and cancelled the user
input. So replaces the `@change` handler with an `@update:model-value`
handler. The `@update:model-value` handler is not called until the user
confirms the input if the `lazy` prop is set.